jsのファイルのディレクトリを分けてリファクタリング
