+++ app/javascript/modules/axios.js
@@ -1,0 +1,6 @@
+import axios from 'axios'
+import { csrfToken } from 'rails-ujs'
+
+axios.defaults.headers.common['X-CSRF-Token'] = csrfToken()
+
+export default axios
