+++ app/javascript/packs/application.js
@@ -24,111 +24,10 @@ axios.defaults.headers.common['X-CSRF-Token'] = csrfToken()
 // const imagePath = (name) => images(name, true)
 // common.js
 
-$(document).on('turbolinks:load',function() {
-  $('.topslide .slide-image:first').addClass('active');
-
-  $('.topslide .arrow.right').click(function() {
-    var active = $('.topslide .slide-image.active');
-    var next = active.next('.slide-image').length ? active.next('.slide-image') : $('.topslide .slide-image:first');
-    active.removeClass('active');
-    next.addClass('active');
-  });
-
-  $('.topslide .arrow.left').click(function() {
-    var active = $('.topslide .slide-image.active');
-    var prev = active.prev('.slide-image').length ? active.prev('.slide-image') : $('.topslide .slide-image:last');
-    active.removeClass('active');
-    prev.addClass('active');
-  });
-});
 
 
 
 
 
 
-document.addEventListener('turbolinks:load', () => {
-  const dataset = $('#article-show').data()
-  const articleId = dataset.articleId
-  axios.get(`/articles/${articleId}/like`)
-  .then((response) => {
-    const hasLiked = response.data.hasLiked
-    if (hasLiked) {
-      $('.like-good').removeClass('hidden')
-  }else{
-    $('.like-nogood').removeClass('hidden')
-  }
- })
-
- $('.like-nogood').on('click', ()=> {
-  axios.post(`/articles/${articleId}/like`)
-  .then((response) => {
-    if (response.data.status === 'ok') {
-      $('.like-nogood').addClass('hidden')
-      $('.like-good').removeClass('hidden')
-    }
-  })
-  .catch((e) => {
-    window.alert('Error')
-    console.log(e)
- })
-})
-
-  $('.like-good').on('click', ()=> {
-    axios.delete(`/articles/${articleId}/like`)
-    .then((response) => {
-      if (response.data.status === 'ok') {
-        $('.like-good').addClass('hidden')
-        $('.like-nogood').removeClass('hidden')
-      }
-    })
-    .catch((e) => {
-      window.alert('Error')
-      console.log(e)
-    })
-  })
-})
-
-
-document.addEventListener("turbolinks:load", function() {
-  let formIdx = document.querySelectorAll(".link").length;
-
-  const wrapper = document.getElementById("links");
-  const addButton = document.querySelector(".add_link");
-  
-  addButton.addEventListener("click", function(e) {  
-    e.preventDefault();
-
-    const newLink = wrapper.querySelector(".link").cloneNode(true);
-    
-    const inputElement = newLink.querySelector("input[type=text]");
-    const nameAttr = inputElement.getAttribute("name");
-    const updatedNameAttr = nameAttr.replace(/\[\d+\]/, `[${formIdx}]`);
-    inputElement.setAttribute("name", updatedNameAttr);
-
-    const regex = /article_links_attributes\[(\d+)\]/g;
-    newLink.innerHTML = newLink.innerHTML.replace(regex, () => {
-      return `article_links_attributes[${formIdx}]`;
-    });
-
-    newLink.querySelector("label").setAttribute("for", `article_article_links_attributes_${formIdx}_url`);
-    newLink.querySelector("input[type=text]").setAttribute("id", `article_article_links_attributes_${formIdx}_url`);
-
-    wrapper.appendChild(newLink); 
-    wrapper.appendChild(addButton);
-    newLink.querySelector(".remove_link").addEventListener("click", removeLink);
-    formIdx++;
-  });
-  
-  function removeLink(e) {  
-    e.preventDefault();  
-    let link = e.target.closest(".link");  
-    link.querySelector("input[type=hidden]").value = "1";  
-    link.style.display = "none";
-  }
-  
-  document.querySelectorAll(".remove_link").forEach(function(removeButton) {  
-    removeButton.addEventListener("click", removeLink);
-  });
-});
