+++ app/javascript/packs/handle_heart.js
@@ -1,0 +1,5 @@
+import $ from 'jquery'
+import axios from 'modules/axios'
+
+
+
