+++ app/views/articles/_form.html.haml
@@ -39,3 +39,4 @@
       %a.add_link.min_btn 追加
     %div
       = f.submit '保存', class: 'btn'
+= javascript_pack_tag 'form'
