+++ app/views/articles/show.html.haml
@@ -60,4 +60,4 @@
       = render 'comments/index', comment: @comment, article: @article
       - if user_signed_in?
         = render 'comments/form', comment: @comment, article: @article
-
+= javascript_pack_tag 'article'
